+++ src/main/java/argo/cost/common/entity/Roles.java
@@ -1,14 +1,20 @@
 package argo.cost.common.entity;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
+import java.sql.Timestamp;
+import java.util.List;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.NamedQuery;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
 
 import org.springframework.context.annotation.Lazy;
-
-import java.sql.Timestamp;
-import java.math.BigDecimal;
-import java.util.List;
 
 
 /**
@@ -24,7 +30,7 @@ public class Roles implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.TABLE)
-	private BigDecimal id;
+	private long id;
 
 	@Column(name="created_date")
 	private Timestamp createdDate;
@@ -53,11 +59,11 @@ public class Roles implements Serializable {
 	public Roles() {
 	}
 
-	public BigDecimal getId() {
+	public long getId() {
 		return this.id;
 	}
 
-	public void setId(BigDecimal id) {
+	public void setId(long id) {
 		this.id = id;
 	}
